+++ crates/gem_xrp/src/typeshare/block.rs
@@ -1,8 +1,10 @@
 use serde::{Deserialize, Serialize};
 use typeshare::typeshare;
 
+use super::UInt64;
+
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[typeshare(swift = "Sendable")]
 pub struct XRPLatestBlock {
-    pub ledger_current_index: i32,
+    pub ledger_current_index: UInt64,
 }
